% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Model fitting}
\usage{
fit_models(data)
}
\arguments{
\item{data}{original sample}
}
\value{
Model estimates
}
\description{
Fit several survival models and return estimates
}
