% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.mc.R
\name{generate.mc}
\alias{generate.mc}
\title{Generate Monte Carlo simulation}
\usage{
generate.mc(num.iter = 10000)
}
\arguments{
\item{num.iter}{number of iterations}
}
\value{
Monte Carlo simulation estimates
}
\description{
Generate 10,000 simulated data set, fit survival models and return estimates
}
