% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bias.R
\name{get_bias}
\alias{get_bias}
\title{Bias of theta}
\usage{
get_bias(results_list, true_theta = log(1))
}
\arguments{
\item{results_list}{list of results}
}
\value{
theta bias
}
\description{
Obtain bias of treatment effect estimates
}
