% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se.R
\name{get_se}
\alias{get_se}
\title{Estimated SE theta}
\usage{
get_se(results_list)
}
\arguments{
\item{results_list}{list of results}
}
\value{
estimated se of theta
}
\description{
Obtain mean of estimated standard error of treatment effect estimates
}
