% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testSize.R
\name{get_testSize}
\alias{get_testSize}
\title{Test size}
\usage{
get_testSize(results_list, true_theta = log(1))
}
\arguments{
\item{results_list}{list of results}
}
\value{
test size
}
\description{
Obtain percentage of simulated data set leading to incorrect rejection of H0
}
