% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theta.R
\name{get_theta}
\alias{get_theta}
\title{Mean theta}
\usage{
get_theta(results_list)
}
\arguments{
\item{results_list}{list of results}
}
\value{
mean theta
}
\description{
Obtain mean of treatment effect estimates
}
