% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data.R
\name{match_data}
\alias{match_data}
\title{Propensity score matching}
\usage{
match_data(o.data)
}
\arguments{
\item{o.data}{original sample}
}
\value{
A 1:1 matched data set
}
\description{
Produce a 1:1 matched data set
}
