% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Data simulation}
\usage{
simulate_data(beta0 = -3.45)
}
\arguments{
\item{beta0}{different levels for covariates imbalance}
}
\value{
A simulated dataset with binary observed covariates, treatment, event time and indicator for 1000 subjects
}
\description{
Simulate a survival data set
}
