% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bias.R, R/get_se.R
\name{get_se}
\alias{get_se}
\title{Bias of theta}
\usage{
get_se(results_list)

get_se(results_list)
}
\arguments{
\item{results_list}{list of results}
}
\value{
theta bias

estimated se of theta
}
\description{
Obtain bias of treatment effect estimates

Obtain mean of estimated standard error of treatment effect estimates
}
